update code 18/4/2025

+++ src/main/java/com/javaweb/Model/BuildingDTO.java
@@ -2,7 +2,7 @@ package com.javaweb.Model;
 // các thực thể
 public class BuildingDTO {
 	private String name;
-	private Integer NOB;
+	private Long NOB;
 	private String adress;
 	private String floorArea;
 
@@ -14,11 +14,11 @@ public class BuildingDTO {
 		this.name = name;
 	}
 
-	public Integer getNOB() {
+	public Long getNOB() {
 		return NOB;
 	}
 
-	public void setNOB(Integer nOB) {
+	public void setNOB(Long nOB) {
 		NOB = nOB;
 	}
